Fix updateProjectInGraph to attempt to write includes after single failure
Before it simply errored out when it hit the first include
that could not be written into neo4j. I've taken the decision
that we want to try the rest of the includes since we might've
just hit some unexpected network issues.

In the case that neo4j has thrown a really bad error this will
continue to throw things at it. But overall the code would
continue anyway just with a different project.

Something to investigate would be if the neo4j driver throws
specific typed errors with which we could assess if we should
continue or not.