Add storage interface and neo4j implementation
The crawler is completely independent from the storage implementation
with this change already. I'm not entirely happy with how the config
for the sotrage implementation is passed through yet, that is not
smooth for the next engine, there is likely a switch case or similar
coming to control what storage is being used, maybe detect it from
the passed configuration instead. It works tho :)