Add an authentication check before trying to stream projects
The GitLab API call for `/projects` always succeeds with a 200 OK if the
instance itself is reachable. To fish out bad credentials early on this
commit implements a new function that tries to call the `/version` endpoint
which required authentication. This will be retried up to 30s.

With this commit I'm also preparing to introduce retries for the rest of
the streaming function at a later point in time. There is an idea in my
head about making a smart choice based on the quality of the connection
i.e. good connections having a faster retry while slow or lossy connections
have a longer slower retry.