Adjust empty project check to not false report
This check should really be moved outside of the main loop.
Running this on every HTTP page request is wasteful, sincce we
only need to check the initial page.

I feel moving back to a less, pretty but more reasonable implementation
of fetching the first page and then go into the loop after might be
worth looking into. The only problem with that is that keyset based
pagination does not give you the pagination header information so it's
a bit less useful, still we can get rid of the if condition in the loop
through this change down the line.