Add potential to parse string sequences into includes
This does complicate the parsing logic a bit, but it seems better
than jumping into reflect and overwriting the YAML parser, it's
simpler and more maintainable in the long run.

Mixed cases are ignored for now since that would almost require
reflection and custom parsers. I feel this catches most of the
common problems that I have experienced for now. I'd rather not
add a component that I am not very familiar or comfortable with
to the architecture at this point.