Add DefaultRefName config option to the crawler
This controls what empty refs are set to, the default reflects
what GitLab is doing actually: https://docs.gitlab.com/ee/ci/yaml/#includefile
> You can also specify a ref. If you do not specify a value, the ref defaults to the HEAD of the project
At this point I don't want to make more calls against the GitLab API
to fetch the current HEAD of each project - but that might be changed
later on. This crawler does make a decent amount of calls as it is
no need to make more for a value that's dynamic and changing constantly
anyway.