Adjust the logging for the new recursive include parsing
The version v0.13.2 has a lot of errors complaining that the code is not
requesting valid file paths and the logging was not disclosing enough
information to track down the actual error.

Now the GitLab client gets handed down a logger to be able to have some
trace logging on the requests, this commit only adds a trace log to the
GetRawFileFromProject requestURL to further debug potential errors.

The error logging during the handling of an input now logs the actual
path of the project it failed during.