Remove projectSet
ProjectSet prevents us from visiting a project more than once,
this is not desirable.

Let's say that we have project B which includes project A's
file f in it's gitlab-ci.yaml. We also have Project A where
it's file f includes project's C. The following diagram will
help illustrate:
B -(file f1)-> A -(file f2)-> C
Order of execution:
A, B, C

If we process project A first we will not discover the
relationship between A and C at first since file f1 in A
includes file f2 in C and we start by processing A's
gitlab-ci.yml, not file f1. since when we visit B we will not
visit A as it has already been visited, we will not traverse
the path from A to C which establishes that A needs C to
execute the gitlab-ci.yml in B

As such we need to explore node A twice, once to parse it's
gitlab-ci.yml and another time as part of the recusrive
resolution of includes of B.

The only chance of an infinite loop being introduced as
part of this change is if a cycle can be formed between file
includes. This will result in an infinite loop of jobs being
created in gitlab so I think it's not possible to have this.