Update module github.com/cenkalti/backoff/v4 to v5
This comes with a tiny rewrite since we need to now always return a
value from our `Operation` function, this was the main breaking change
between v4 and v5, so instead of returning only an error we return an
empty struct that is promptly ignored. This should come with no
additional memory cost since empty structs are of size 0.

Also bumps the Golang version to 1.24 because it was a nightmare to do
outside of this change and I was lazy.